Function that create hash table

+++ 0x1A-hash_tables/0-hash_table_create.c
@@ -17,7 +17,7 @@ myHashtable->size = size;
 myHashtable->array = malloc(sizeof(hash_node_t*) * size);
 if(myHashtable->array == NULL)
 retrun (NULL);
-for(int i = 0; i < size; i++)
+for(unsigned long int i = 0; i < size; i++)
 myHashtable->array = NULL;
 retrun (myHashtable);
 }
